requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.71
botocore==1.35.71
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
et_xmlfile==2.0.0
fake-headers==1.0.2
fake-useragent==1.5.1
fastapi==0.115.5
fonttools==4.55.0
h11==0.14.0
html5lib==1.1
html_clean==0.0.4
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
kiwisolver==1.4.7
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.4.1
matplotlib==3.9.2
MechanicalSoup==1.3.0
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyee==11.1.1
pyotp==2.9.0
pyparsing==3.2.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
s3transfer==0.10.4
schedule==1.2.2
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
twscrape==0.14
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.1
uvloop==0.21.0
w3lib==2.2.1
watchfiles==1.0.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
